feat: add consent-skip on sign-up for trusted clients (PDS_SIGNUP_ALLOW_CONSENT_SKIP)
Move resolveClientMetadata to shared package so pds-core can resolve
custom ePDS metadata fields. Add triple-gated consent-skip: env var +
trusted client + client metadata flag. When not skipping, new users
get prompt=consent to bypass unnecessary account selection screen.

Also adds AGENTS.md debugging guidance.
