+++ .env.example
@@ -85,6 +85,14 @@ PDS_EMAIL_SMTP_URL=smtp://localhost:1025
 PDS_EMAIL_FROM_ADDRESS=noreply@pds.example
 PDS_BLOBSTORE_DISK_LOCATION=/data/blobs
 
+# Trusted OAuth clients — comma-separated client_id URLs.
+# PDS_OAUTH_TRUSTED_CLIENTS=https://app.example/client-metadata.json
+
+# Skip consent on sign-up for trusted clients that request it.
+# Requires PDS_OAUTH_TRUSTED_CLIENTS and client metadata with
+# "epds_skip_consent_on_signup": true. Default: false.
+# PDS_SIGNUP_ALLOW_CONSENT_SKIP=false
+
 # Invite code for automated account creation (ePDS creates accounts on first login).
 # Required when PDS_INVITE_REQUIRED is true (the default).
 # Generate with:
