+++ packages/pds-core/.env.example
@@ -81,5 +81,15 @@ PDS_EMAIL_FROM_ADDRESS=noreply@pds.example
 # PDS_MODERATION_EMAIL_SMTP_URL=
 # PDS_MODERATION_EMAIL_ADDRESS=
 
+# Trusted OAuth clients (comma-separated client_id URLs).
+# Trusted clients get relaxed consent handling and may inject custom branding.
+# PDS_OAUTH_TRUSTED_CLIENTS=https://app.example/client-metadata.json
+
+# When true, consent may be skipped on initial sign-up for trusted clients
+# whose metadata includes "epds_skip_consent_on_signup": true.
+# All three conditions must be met: this env var is truthy, the client is in
+# PDS_OAUTH_TRUSTED_CLIENTS, and the client's metadata requests the skip.
+# PDS_SIGNUP_ALLOW_CONSENT_SKIP=false
+
 # Blobstore — where to store uploaded blobs (images, etc.)
 PDS_BLOBSTORE_DISK_LOCATION=/data/blobs
